+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * Added a call to exit users from A/B experiment with experiment ID : `testingExitABExperiment:`
 
 * Fixed the exit AB request failure issue on iOS.
-* Fixed "null" value issue for experiment info variants.
+* Fixed 'null' value issue for experiment info variants.
+* Fixed 'null' pointer error thrown when 'testingGetVariantsForKey' method is called with a key that does not exist.
 
 * Updated underlying Android SDK version to 23.8.3
 * Underlying iOS SDK version is 23.8.3
@@ -13,7 +14,8 @@
 * Added a call to exit users from A/B experiment with experiment ID : `testingExitABExperiment:`
 
 * Fixed the exit AB request failure issue on iOS.
-* Fixed "null" value issue for experiment info variants.
+* Fixed 'null' value issue for experiment info variants.
+* Fixed 'null' pointer error thrown when 'testingGetVariantsForKey' method is called with a key that does not exist.
 
 * Updated underlying Android SDK version to 23.8.3
 * Underlying iOS SDK version is 23.8.3